chore: move platform and arch utils to own module
