+++ core/src/commands/self-update.ts
@@ -14,8 +14,9 @@ import type { GlobalOptions, ParameterValues } from "../cli/params.js"
 import { BooleanParameter, ChoicesParameter, StringParameter } from "../cli/params.js"
 import { dedent } from "../util/string.js"
 import { basename, dirname, join, resolve } from "path"
-import type { Architecture } from "../util/util.js"
-import { getArchitecture, isDarwinARM, getPackageVersion, getPlatform } from "../util/util.js"
+import type { Architecture } from "../util/arch-platform.js"
+import { getArchitecture, isDarwinARM, getPlatform } from "../util/arch-platform.js"
+import { getPackageVersion } from "../util/util.js"
 import { RuntimeError } from "../exceptions.js"
 import { makeTempDir } from "../util/fs.js"
 import { createReadStream, createWriteStream } from "fs"
