+++ core/src/util/ext-tools.ts
@@ -10,7 +10,8 @@ import fsExtra from "fs-extra"
 const { pathExists, createWriteStream, ensureDir, chmod, remove, move, createReadStream } = fsExtra
 import { ConfigurationError, InternalError } from "../exceptions.js"
 import { join, dirname, basename, posix } from "path"
-import { hashString, exec, getPlatform, getArchitecture, isDarwinARM } from "./util.js"
+import { getArchitecture, getPlatform, isDarwinARM } from "./arch-platform.js"
+import { hashString, exec } from "./util.js"
 import tar from "tar"
 import { GARDEN_GLOBAL_PATH } from "../constants.js"
 import type { Log } from "../logger/log-entry.js"
