+++ core/src/util/util.ts
@@ -7,7 +7,6 @@
  */
 
 import { asyncExitHook, gracefulExit } from "@scg82/exit-hook"
-import { execSync } from "child_process"
 import _spawn from "cross-spawn"
 import { createHash } from "node:crypto"
 import fsExtra from "fs-extra"
@@ -21,7 +20,6 @@ import {
   isArray,
   isPlainObject,
   mapValues,
-  memoize,
   omit,
   pick,
   range,
@@ -39,10 +37,10 @@ import { DEFAULT_GARDEN_CLOUD_DOMAIN, DOCS_BASE_URL, gardenEnv } from "../consta
 import {
   ChildProcessError,
   InternalError,
+  isErrnoException,
   ParameterError,
   RuntimeError,
   TimeoutError,
-  isErrnoException,
 } from "../exceptions.js"
 import type { Log } from "../logger/log-entry.js"
 import { getDefaultProfiler } from "./profiling.js"
@@ -195,6 +193,7 @@ export function createOutputStream(log: Log, origin?: string) {
 
   return outputStream
 }
+
 export interface ExecOpts extends ExecaOptions {
   stdout?: Writable
   stderr?: Writable
@@ -597,57 +596,6 @@ export function isSubdir(path: string, ofPath: string): boolean {
   const rel = relative(path, ofPath)
   return !!(rel && !rel.startsWith("..") && !isAbsolute(rel))
 }
-
-// Used to make the platforms more consistent with other tools
-const platformMap = {
-  win32: "windows" as const,
-}
-
-const archMap = {
-  x32: "386" as const,
-  x64: "amd64" as const,
-}
-
-export type Architecture = Exclude<NodeJS.Architecture, keyof typeof archMap> | (typeof archMap)[keyof typeof archMap]
-export type Platform =
-  | Exclude<NodeJS.Platform, keyof typeof platformMap>
-  | (typeof platformMap)[keyof typeof platformMap]
-
-export function getPlatform(): Platform {
-  return platformMap[process.platform] || process.platform
-}
-
-export function getArchitecture(): Architecture {
-  // Note: When node is running a x64 build,
-  // process.arch is always x64 even though the underlying CPU architecture may be arm64
-  // To check if we are running under Rosetta,
-  // use the `isDarwinARM` function below
-  const arch = process.arch
-  return archMap[arch] || arch
-}
-
-export const isDarwinARM = memoize(() => {
-  if (process.platform !== "darwin") {
-    return false
-  }
-
-  if (process.arch === "arm64") {
-    return true
-  } else if (process.arch === "x64") {
-    // detect rosetta on Apple M cpu family macs
-    // see also https://developer.apple.com/documentation/apple-silicon/about-the-rosetta-translation-environment
-    // We use execSync here, because this function is called in a constructor
-    // otherwise we'd make the function async and call `spawn`
-    try {
-      execSync("sysctl -n -q sysctl.proc_translated", { encoding: "utf-8" })
-    } catch (err) {
-      return false
-    }
-    return true
-  }
-
-  return false
-})
 
 export function getDurationMsec(start: Date, end: Date): number {
   return Math.round(end.getTime() - start.getTime())
